server/requirements.txt
mediapipe==0.10.9
numba==0.59.0
numpy==2.1.2
opencv_contrib_python==4.9.0.80
opencv_python==4.9.0.80
python_bcrypt==0.3.2
torch==2.2.0
websockets==12.0
